started permissions

+++ src/app/Http/Controllers/User/UserController.php
@@ -18,7 +18,8 @@ use Illuminate\Http\Request;
 class UserController extends Controller {
 
     public function showProfile(string $username) {
-	    if (!Auth::check()) return redirect('/login');
+        if (!Auth::check()) return redirect('/login');
+        
         $this->authorize('list', Project::class);
 
         try {
@@ -34,10 +35,9 @@ class UserController extends Controller {
             'sprintsContributedTo' => $sprintsContributedTo, 'user' => $user]);
 
         } catch(\Illuminate\Database\QueryException $qe) {
-            // Catch the specific exception and handle it 
-            //(returning the view with the parsed errors, p.e)
+            return redirect()->route('error');
         } catch (\Exception $e) {
-            // Handle unexpected errors
+            return redirect()->route('error');
         }
 
       
@@ -49,10 +49,9 @@ class UserController extends Controller {
             return view('partials.user_projects', ['projects' => $projects]);
 
         } catch(\Illuminate\Database\QueryException $qe) {
-            // Catch the specific exception and handle it 
-            //(returning the view with the parsed errors, p.e)
+            return redirect()->route('error');
         } catch (\Exception $e) {
-            // Handle unexpected errors
+            return redirect()->route('error');
         }
         
     }
@@ -60,7 +59,10 @@ class UserController extends Controller {
 	public function showAdminPage(string $username){
 		if (!Auth::check()) return redirect('/login');
 
-		return view('pages/admin_page');
+        if(Auth::user()->username === $username)
+            return view('pages/admin_page');
+        else
+            return redirect()->route('error');
 	}
 
     /**]);
@@ -85,10 +87,9 @@ class UserController extends Controller {
             return response()->json(array('success' => true, 'html' => $viewHTML));
 
         } catch(\Illuminate\Database\QueryException $qe) {
-            // Catch the specific exception and handle it 
-            //(returning the view with the parsed errors, p.e)
+            return redirect()->route('error');
         } catch (\Exception $e) {
-            // Handle unexpected errors
+            return redirect()->route('error');
         }
             
     }
@@ -98,27 +99,33 @@ class UserController extends Controller {
 	*/
 	public function editProfileAction(Request $request) {
         if (!Auth::check()) return redirect('/login');
-        $user = Auth::user();
+        
         try {
-            $user->name = $request->user_name;
-            $user->username = $request->user_username;
-            $user->email = $request->user_email;
-            
-            if($request->hasFile('user_image')){
-                $file = $request->file('user_image');
-                echo $file;
-                $user->image = $request->file('user_image')->store('');              
+            if(Auth::user()->username === $request->username){
+                $user = Auth::user();
+    
+                $user->name = $request->user_name;
+                $user->username = $request->user_username;
+                $user->email = $request->user_email;
+                
+                if($request->hasFile('user_image')){
+                    $file = $request->file('user_image');
+                    echo $file;
+                    $user->image = $request->file('user_image')->store('');              
+                }
+        
+        
+                $user->save();
+                return back();
             }
-    
-    
-            $user->save();
-            return back();
+            else {
+                return redirect()->route('error');
+            }
 
         } catch(\Illuminate\Database\QueryException $qe) {
-            // Catch the specific exception and handle it 
-            //(returning the view with the parsed errors, p.e)
+            return redirect()->route('error');
         } catch (\Exception $e) {
-            // Handle unexpected errors
+            return redirect()->route('error');
         }
     }
     
@@ -129,17 +136,20 @@ class UserController extends Controller {
         if (!Auth::check()) return redirect('/login');
 
         try {
-
             $notification = Notification::find($notification_id);
-            $notification->delete();
-
-            return response()->json(array('success' => true, 'notification_id' => $notification_id));
+            if(Auth::user()->id === $notification->user_id){
+                $notification->delete();
+    
+                return response()->json(array('success' => true, 'notification_id' => $notification_id));
+            }
+            else {
+                return redirect()->route('error');
+            }
             
         } catch(\Illuminate\Database\QueryException $qe) {
-            // Catch the specific exception and handle it 
-            //(returning the view with the parsed errors, p.e)
+            return redirect()->route('error');
         } catch (\Exception $e) {
-            // Handle unexpected errors
+            return redirect()->route('error');
         }
     }
 
@@ -165,10 +175,9 @@ class UserController extends Controller {
             return response()->json(array('success' => true, 'notification_id' => $notification_id));
             
         } catch(\Illuminate\Database\QueryException $qe) {
-            // Catch the specific exception and handle it 
-            //(returning the view with the parsed errors, p.e)
+            return redirect()->route('error');
         } catch (\Exception $e) {
-            // Handle unexpected errors
+            return redirect()->route('error');
         } 
     }
 
@@ -190,10 +199,9 @@ class UserController extends Controller {
             return response()->json(array('success' => true, 'notification_id' => $notification_id));
             
         } catch(\Illuminate\Database\QueryException $qe) {
-            // Catch the specific exception and handle it 
-            //(returning the view with the parsed errors, p.e)
+            return redirect()->route('error');
         } catch (\Exception $e) {
-            // Handle unexpected errors
+            return redirect()->route('error');
         } 
     }
 
@@ -224,10 +232,9 @@ class UserController extends Controller {
             }            
             
         } catch(\Illuminate\Database\QueryException $qe) {
-            // Catch the specific exception and handle it 
-            //(returning the view with the parsed errors, p.e)
+            return redirect()->route('error');
         } catch (\Exception $e) {
-            // Handle unexpected errors
+            return redirect()->route('error');
         }
     }
 
@@ -240,10 +247,9 @@ class UserController extends Controller {
             return response()->json(array('success' => true,'html' => $html));
 
         }catch(\Illuminate\Database\QueryException $qe) {
-            // Catch the specific exception and handle it 
-            //(returning the view with the parsed errors, p.e)
+            return redirect()->route('error');
         } catch (\Exception $e) {
-            // Handle unexpected errors
+            return redirect()->route('error');
         }
     }
 
@@ -256,10 +262,9 @@ class UserController extends Controller {
             return response()->json(array('success' => true,'html' => $html));
 
         }catch(\Illuminate\Database\QueryException $qe) {
-            // Catch the specific exception and handle it 
-            //(returning the view with the parsed errors, p.e)
+            return redirect()->route('error');
         } catch (\Exception $e) {
-            // Handle unexpected errors
+            return redirect()->route('error');
         }
     }
 }
